Move loading-instructions into constants
