+++ src/preact/context.js
@@ -15,7 +15,7 @@
  */
 
 import * as Preact from '#preact';
-import {Loading, reducer as loadingReducer} from '#core/loading-instructions';
+import {Loading, reducer as loadingReducer} from '#core/constants/loading-instructions';
 import {createContext, useContext, useMemo} from '#preact';
 
 /** @type {PreactDef.Context} */
